backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-decouple==3.8
pydantic==2.10.4
pydantic-settings==2.1.0
asyncpg==0.30.0
databases[postgresql]==0.9.0
fastapi-users==13.0.0
fastapi-users[sqlalchemy]==13.0.0
websockets==14.1
aiofiles==24.1.0
Pillow==11.3.0
typing-extensions==4.14.1